requirements.txt
streamlit==1.38.0
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-chroma==0.1.3
langchain-huggingface==0.0.3
langchain-groq==0.1.9
nltk==3.8.1
pypdf==3.17.4
pdfplumber==0.10.3
chromadb==0.4.24
tqdm==4.66.4
python-dotenv==1.0.1
